==> Pandas/1 - Series.ipynb <==
# Series - Pandas


import pandas as pd

dados = [10,20,30,40,50]
serie = pd.Series(dados)
serie

type(serie)

# 1 - Criando uma Series com dados de vendas mensais
vendas_mensais = pd.Series([2000,2500,3000,3500,4000,4500,5000,4800,5200,6000,7000,8000],
                           index=['jan', 'fev', 'mar', 'abr', 'mai', 'jun', 'jul', 'ago', 'set', 'out', 'nov', 'dez'])

vendas_mensais

# Calcular e exibir a média de vendas mensais
media_vendas = vendas_mensais.mean()
print(media_vendas)

#3 - Identificando o mês com maior venda
mes_maior_venda = vendas_mensais.idxmax()
mes_maior_venda

# exibir vendas acima da média
vendas_acima_media = vendas_mensais[vendas_mensais > media_vendas]
vendas_acima_media
