==> Pandas/2 - Dataframe.ipynb <==
# Dataframe


import pandas as pd

dados = {
    'Nome': ['Aline', 'Bob', 'Charlie'],
    'Idade': [24,30,22],
    'Cidade': ['São Paulo', 'Rio de Janeiro', 'Belo Horizonte']
}

df = pd.DataFrame(dados)
df

df.head(1)

df.tail(1) # traz últtima linha

dados_notas = {
    'Nome':['Aline', 'Bob', 'Charlie', 'David', 'Eva' ],
    'Matematica':[85,90,78,88,92],
    'Fisica': [80,75,88,85,95],
    'Química': [78,82,85,80,88]
    }

df_notas = pd.DataFrame(dados_notas)
df_notas

# Calculando a média das notas
df_notas['Media'] = df_notas[['Matematica', 'Fisica', 'Química']].mean(axis=1).round(2)
df_notas

# Identificar alunos aprovados com nota boa (média >= 82.5)
alunos_aprovados = df_notas[df_notas['Media']>=82.5]
alunos_aprovados

# Exibir a média e os alunos aprovados

df_notas[['Nome', 'Media']]

alunos_aprovados[['Nome', 'Media']]
