==> Pandas/5 - Importando_CSV.ipynb <==
# Importando arquivos CSV

import pandas as pd

caminho = 'C:\\Users\\Matheus\\Documents\\Analise_de_dados\\Dados\\produtos.csv'
df_produtos = pd.read_csv(caminho, sep = ',')

df_produtos.head(8)

df_produtos.tail(8)

# Verifica as dimensões do Dataframe (linhas, colunas)
df_produtos.shape

# Verificando os tipos das variáveis
df_produtos.dtypes

df_produtos.sort_values(by='Quantidade', ascending=False)

# Verificando a quantidade de produtos pela categoria 
df_produtos['Categoria'].value_counts()
