==> Pandas/6 - insights.ipynb <==
# Tirando insights do dataframe

import pandas as pd

caminho = '..\\Dados\\produtos.csv' # .. (volta um diretório anterior pois a pasta Dados está localizada em outra pasta)
df_produtos = pd.read_csv(caminho)

df_produtos

# Verificando os tipos das variáveis
df_produtos.dtypes

df_produtos['Quantidade'] = pd.to_numeric(df_produtos['Quantidade'])
df_produtos['Preço'] = pd.to_numeric(df_produtos['Preço'])

#1 - Desempenho de Vendas por Categoria

vendas_categoria = df_produtos.groupby('Categoria')['Quantidade'].sum()
vendas_categoria

# Produto mais vendido
produto_mais_vendido = df_produtos.groupby('Produto')['Quantidade'].sum().sort_values(ascending=False).head(2)
produto_mais_vendido

# Preço médio por produto 
media_preco_produto = df_produtos.groupby('Produto')['Preço'].mean().sort_values(ascending=False)
media_preco_produto

# Correlação entre unidades vendidas e preço unitário
correlacao = df_produtos['Quantidade'].corr(df_produtos['Preço'])
print(correlacao)

vendas_categoria_df = vendas_categoria.reset_index()
vendas_categoria_df.columns = ['Categoria', 'UnidadesVendidas']
vendas_categoria_df

vendas_categoria_df.to_csv('..\\dados\\vendas_categoria.csv', index=False)
