==> Pandas/8- Analise_planilha.ipynb <==
## Análise de dados direto de planilha 

import pandas as pd
import openpyxl

df = pd.read_excel('../Dados/campeonato_brasileiro.xlsx')

df.head()

# Verificando os tipos de dados
df.dtypes

# Medidas estatísticas
df.describe()

# primeiros 2 colocados
df.head(2)

# últimos 2 colocados
df.tail(2)

# 1 - Analise desempenho por time
desempenho_info = df[['Times', 'Pontos', 'Vitorias', 'Saldo_Gols']].set_index('Times')
desempenho_info

# Time com mais vitórias
time_mais_vitorias = df[df['Vitorias']==df['Vitorias'].max()]
time_mais_vitorias

# Time menos vitórias
time_menos_vitorias = df[df['Vitorias']==df['Vitorias'].min()]
time_menos_vitorias


# Distribuição de gols pró
df['Gols_Pro'].value_counts()

# relação entre gols pró e pontos 
correlação = df['Gols_Pro'].corr(df['Pontos'])
correlação


correlação = df['Gols_Contra'].corr(df['Pontos'])
correlação
