==> Pandas/9 - import_json.ipynb <==
# Trabalhando com Arquivos Json

import json

dados_receitas = {
    'receitas': [
        {
            'titulo': 'Bolo de chocolate',
            'tempo_preparo': '45 minutos',
            'porcoes': 8,
            'ingredientes':[
                {'ingrediente': 'Farinha de trigo', 'quantidade': '2 xicaras'},
                {'ingrediente': 'Açúcar', 'quantidade':'1 xícara'},
                {'ingrediente': 'Ovos', 'Quantidade': '3 xícaras'},
                {'ingrediente': 'Leite', 'Quantidade': '1 xícara'},
                {'ingrediente': 'Manteiga', 'Quantidade': '1/2 xícaras'}
            ]
        },
        {
            'titulo': 'Salada Caesar',
            'tempo_preparo': '20 minutos',
            'porcoes': 4,
            'ingredientes': [
                {'ingrediente':'Alface romana', 'quantidade': '1 cabeça'},
                {'ingrediente':'Croutons', 'quantidade': '1 xícara'},
                {'ingrediente':'Queijo Parmesão', 'quantidade': '1/2 xícara'},
                {'ingrediente':'Molho Caesar', 'quantidade': '1/4 xícara'}, 
            ]
        }
    ]
}

with open ('../Dados/receitas.json', 'w', encoding='utf-8') as file:
    json.dump(dados_receitas, file, ensure_ascii=False, indent=4)
