==> Pandas/11-Analisando _json.ipynb <==
# Analisando dados de arquivo Json

import pandas as pd



df = pd.read_json('../Dados/receitas.json', encoding='utf-8')
df

df_receitas = pd.json_normalize(df['receitas'],
                            'ingredientes',   
                            ['titulo', 'tempo_preparo', 'porcoes'])
df_receitas

# Análise Total de Ingredientes por receita
receitas_expandidas = []

for receita in df['receitas']:
    total_ingredientes = (len(receita['ingredientes']))
    receita_info = {
        'titulo':receita['titulo'],
        'tempo_preparo': receita['tempo_preparo'],
        'porcoes':receita['porcoes'],
        'total_ingredientes': total_ingredientes

    }
    receitas_expandidas.append(receita_info)

receitas_expandidas

df_expandido = pd.DataFrame(receitas_expandidas)
df_expandido
