==> Pandas/12-bd.ipynb <==
## Transformando um arquivo CSV em tabela de banco de dados

import pandas as pd
from sqlalchemy import create_engine

# Criando o arquivo do banco na pasta Dados
engine = create_engine('sqlite:///../Dados/ecommerce.db')

df = pd.read_csv("../Dados/produtos.csv")

df

try:
    df.to_sql('produtos', con=engine, if_exists='replace', index=False)
except Exception as e:
    print(f'Ocorreu um erro ao exportar os dados {e}')
