==> Pandas/13-analise_bd.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

# Conectando ao banco de dados
engine = create_engine('sqlite:///../Dados/ecommerce.db')

# Testando conexão do banco de dados
with engine.connect() as connection:
    print('Conectado com sucesso')

# 1 - Total de vendas por categoria
consulta_1 = 'SELECT Categoria, sum(Quantidade) AS Total_QTD_Vendida FROM produtos GROUP BY Categoria'
resultado_1 = pd.read_sql_query(consulta_1, engine)
resultado_1

# Produto com mais vendas
consulta_2 = 'SELECT Produto, SUM(Quantidade) AS Total_Qtd_Vendida FROM produtos GROUP BY Produto ORDER BY Total_Qtd_Vendida DESC LIMIT 1'
resultado_2 = pd.read_sql_query(consulta_2, engine)
resultado_2

# Total de vendas por produto
consulta_3 = 'SELECT Produto, SUM(Quantidade) as Total_Qtd_Vendida FROM produtos GROUP BY Produto ORDER BY Total_Qtd_vendida'
resultado_3 = pd.read_sql_query(consulta_3, engine)
resultado_3

# 4 média de preço por produto
consulta_4 = 'SELECT Produto, AVG(Preço) AS Media_Preco FROM produtos GROUP BY Produto'
resultado_4 = pd.read_sql_query(consulta_4, engine)
resultado_4
